Move Core Data `Managed Objects` to it's own file -> `ManagedNewsItem`
